# src/main/java/com/ai/comics/api/AnalyticsApi.java
```java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ai.comics.api;

import com.ai.comics.model.Analytics;
import com.ai.comics.model.AnalyticsResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-01-13T13:20:51.374101Z[Etc/UTC]", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "analytics", description = "the analytics API")
public interface AnalyticsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /analytics/{analyticsId} : Delete an analytics record by ID
     *
     * @param analyticsId  (required)
     * @return Analytics record deleted successfully (status code 204)
     */
    @Operation(
        operationId = "analyticsAnalyticsIdDelete",
        summary = "Delete an analytics record by ID",
        responses = {
            @ApiResponse(responseCode = "204", description = "Analytics record deleted successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/analytics/{analyticsId}"
    )
    
    default ResponseEntity<Void> analyticsAnalyticsIdDelete(
        @Parameter(name = "analyticsId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("analyticsId") Integer analyticsId
    ) throws Exception {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /analytics/{analyticsId} : Get an analytics record by ID
     *
     * @param analyticsId  (required)
     * @return A single analytics record (status code 200)
     */
    @Operation(
        operationId = "analyticsAnalyticsIdGet",
        summary = "Get an analytics record by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "A single analytics record", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AnalyticsResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/analytics/{analyticsId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AnalyticsResponse> analyticsAnalyticsIdGet(
        @Parameter(name = "analyticsId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("analyticsId") Integer analyticsId
    ) throws Exception {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"shares\" : 5, \"comic_id\" : 6, \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"analytics_id\" : 0, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"views\" : 1, \"likes\" : 5 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /analytics/{analyticsId} : Update an analytics record by ID
     *
     * @param analyticsId  (required)
     * @param analytics  (required)
     * @return Analytics record updated successfully (status code 200)
     */
    @Operation(
        operationId = "analyticsAnalyticsIdPut",
        summary = "Update an analytics record by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Analytics record updated successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/analytics/{analyticsId}",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> analyticsAnalyticsIdPut(
        @Parameter(name = "analyticsId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("analyticsId") Integer analyticsId,
        @Parameter(name = "Analytics", description = "", required = true) @Valid @RequestBody Analytics analytics
    ) throws Exception {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /analytics : Get all analytics records
     *
     * @return A list of analytics records (status code 200)
     */
    @Operation(
        operationId = "analyticsGet",
        summary = "Get all analytics records",
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of analytics records", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = AnalyticsResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/analytics",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<AnalyticsResponse>> analyticsGet(
        
    ) throws Exception {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"shares\" : 5, \"comic_id\" : 6, \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"analytics_id\" : 0, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"views\" : 1, \"likes\" : 5 }, { \"shares\" : 5, \"comic_id\" : 6, \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"analytics_id\" : 0, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"views\" : 1, \"likes\" : 5 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /analytics : Create a new analytics record
     *
     * @param analytics  (required)
     * @return Analytics record created successfully (status code 201)
     */
    @Operation(
        operationId = "analyticsPost",
        summary = "Create a new analytics record",
        responses = {
            @ApiResponse(responseCode = "201", description = "Analytics record created successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/analytics",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> analyticsPost(
        @Parameter(name = "Analytics", description = "", required = true) @Valid @RequestBody Analytics analytics
    ) throws Exception {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}
```